@@ -13,25 +13,25 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@tauri-apps/api": "^2",
-    "@tauri-apps/plugin-log": "~2",
-    "@tauri-apps/plugin-opener": "^2",
-    "@tauri-apps/plugin-sql": "~2"
+    "@tauri-apps/api": "^2.4.1",
+    "@tauri-apps/plugin-log": "~2.3.1",
+    "@tauri-apps/plugin-opener": "^2.2.6",
+    "@tauri-apps/plugin-sql": "~2.2.0"
   },
   "devDependencies": {
-    "@sveltejs/adapter-static": "^3.0.6",
-    "@sveltejs/kit": "^2.9.0",
-    "@sveltejs/vite-plugin-svelte": "^5.0.0",
-    "@tailwindcss/vite": "^4.0.14",
-    "@tauri-apps/cli": "^2",
+    "@sveltejs/adapter-static": "^3.0.8",
+    "@sveltejs/kit": "^2.20.4",
+    "@sveltejs/vite-plugin-svelte": "^5.0.3",
+    "@tailwindcss/vite": "^4.1.3",
+    "@tauri-apps/cli": "^2.4.1",
     "date-fns": "^4.1.0",
     "flowbite": "^3.1.2",
-    "flowbite-svelte": "^0.48.4",
-    "flowbite-svelte-icons": "^2.0.3",
-    "svelte": "^5.0.0",
-    "svelte-check": "^4.0.0",
-    "tailwindcss": "^4.0.14",
-    "typescript": "~5.6.2",
-    "vite": "^6.0.3"
+    "flowbite-svelte": "^0.48.5",
+    "flowbite-svelte-icons": "^2.1.1",
+    "svelte": "^5.25.8",
+    "svelte-check": "^4.1.5",
+    "tailwindcss": "^4.1.3",
+    "typescript": "~5.6.3",
+    "vite": "^6.2.5"
   }
 }
