@@ -128,23 +128,41 @@ pub struct NewTodoList {
 #[diesel(table_name = todos)]
 pub struct Todo {
     pub id: i32,
-    pub list_id: i32,
+    pub calendar_id: i32,
     pub uid: String,
-    pub etag: String,
-    pub url: String,
+    pub href: String,
     pub ical_data: String,
-    pub last_modified: i64,
+    pub summary: String,
+    pub description: Option<String>,
     pub completed: bool,
+    pub tag: Option<String>,
+    pub status: EventStatus,
+    pub event_type: EventType,
+    pub original_text: Option<String>,
+    pub load: i32,
+    pub urgency: i32,
+    pub importance: i32,
+    pub postponed: i32,
+    pub last_modified: i64,
 }
 
 #[derive(Queryable, Selectable, Insertable, AsChangeset, Debug)]
 #[diesel(table_name = todos)]
 pub struct NewTodo {
-    pub list_id: i32,
+    pub calendar_id: i32,
+    pub href: String,
     pub uid: String,
-    pub etag: String,
-    pub url: String,
     pub ical_data: String,
-    pub last_modified: i64,
+    pub summary: String,
+    pub description: Option<String>,
     pub completed: bool,
+    pub tag: Option<String>,
+    pub status: EventStatus,
+    pub event_type: EventType,
+    pub original_text: Option<String>,
+    pub load: i32,
+    pub urgency: i32,
+    pub importance: i32,
+    pub postponed: i32,
+    pub last_modified: i64,
 }
